requirements.txt
pytest==6.2.4
SQLAlchemy==1.4.21
grpcio==1.39.0
PyYAML==5.4.1
protobuf==3.17.3
